requirements.txt
#Deepeval
deepeval

#ragas
ragas
tqdm

#AI
openai
google-generativeai
anthropic
langchain
langchain-openai


#UI
streamlit

pandas
datasets
rouge-score
scikit-learn
nltk
matplotlib
seaborn
python-dotenv
absl-py
evaluate
spacy
dotenv
python-dotenv
javalang
